Added Hardware SPI functionality to soft_spi
The now misnamed soft_spi now has the ability to utilize the hardware
spi code from the last commit. This can be used with little changes to
the other code in this project, and can support baud rates up to
25MBaud. This was tested on my personal DE0 with an analog discovery
using the unit test file. The write functionality is definitely working,
and was able to perform an entire getQuatRates call in only 250us at
5MBaud, down from around 1100us with bit banging.