Fixed ImageCaptureWrapper error and added capture routine
Th image capture wrapper has been tested with echo functions in place
of the actual ZWO calls, but appears to be fully functional. A
potential Image capture routine was implemented in the python
controller, but this will need to be tested. Also allowed for spaces
before command line entries to try and fix the command rejection issue.